Parameterized output type for the `uml` target.
By default, the `make uml` rule now outputs a .dot file, which is
automatically converted to a .graphml file. This can be overridden by
setting the `UML_IMG_TYPE` variable when calling `make uml`. For
example, if you wanted to save the diagrams as PNG files then you would
run

    `make UML_IMG_TYPE=png uml`